build_python_producer/requirement.txt
Flask
kafka-python
six
